refactor for python code style changes
