+++ Dockerfile
@@ -7,6 +7,7 @@ FROM quay.io/centos/centos:stream8
 RUN dnf module -y install python39 && dnf install -y python39 python39-pip lksctp-tools-devel
 COPY --from=builder /usr/local/bin/uperf /usr/local/bin/uperf
 RUN mkdir /plugin
+ADD https://raw.githubusercontent.com/arcalot/arcaflow-plugins/main/LICENSE /plugin
 ADD uperf_plugin.py /plugin
 ADD uperf_schema.py /plugin
 ADD test_uperf_plugin.py /plugin
@@ -21,7 +22,7 @@ EXPOSE 20000
 ENTRYPOINT ["python3.9", "uperf_plugin.py"]
 CMD []
 
-LABEL org.opencontainers.image.source="https://github.com/arcalot/arcaflow-plugins/blob/main/python/uperf/Dockerfile"
+LABEL org.opencontainers.image.source="https://github.com/arcalot/arcaflow-plugin-uperf"
 LABEL org.opencontainers.image.licenses="Apache-2.0+GPL-3.0-only"
 LABEL org.opencontainers.image.vendor="Arcalot project"
 LABEL org.opencontainers.image.authors="Arcalot contributors"
