+++ input/connect.yaml
