--- python/_example_plugin/README.md
@@ -1,13 +1,0 @@
-# Example Arcaflow plugin
-
-This is an example plugin for Arcaflow that just says hello. You can use it by creating the following configuration file:
-
-```yaml
-name: Your Name
-```
-
-You can run the plugin by typing:
-
-```
-./example_plugin.py -f your-config.yaml
-```
