--- python/_example_plugin/requirements.txt
@@ -1,1 +1,0 @@
-arcaflow_plugin_sdk
