+++ test_uperf_plugin.py
@@ -8,49 +8,40 @@ from arcaflow_plugin_sdk import plugin
 
 
 simple_profile = uperf_schema.Profile(
-    name="test", groups=[
+    name="test",
+    groups=[
         uperf_schema.ProfileGroup(
             nthreads=1,
-            transactions= [
+            transactions=[
                 uperf_schema.ProfileTransaction(
-                    iterations = 1,
-                    flowops = [
+                    iterations=1,
+                    flowops=[
                         uperf_schema.AcceptFlowOp(
                             type="accept",
                             remotehost="127.0.0.1",
                             protocol=uperf_schema.IProtocol.TCP,
                             wndsz=5,
-                            tcp_nodelay=True
+                            tcp_nodelay=True,
                         )
-                    ]
+                    ],
                 ),
                 uperf_schema.ProfileTransaction(
                     duration="50ms",
-                    flowops = [
-                        uperf_schema.WriteFlowOp(
-                            type="write",
-                            size=90
-                        ),
-                        uperf_schema.ReadFlowOp(
-                            type="read",
-                            size=90
-                        )
-                    ]
+                    flowops=[
+                        uperf_schema.WriteFlowOp(type="write", size=90),
+                        uperf_schema.ReadFlowOp(type="read", size=90),
+                    ],
                 ),
                 uperf_schema.ProfileTransaction(
-                    iterations = 1,
-                    flowops = [
-                        uperf_schema.DisconnectFlowOp(
-                            type="disconnect"
-                        )
-                    ]
-                )
-            ]
+                    iterations=1,
+                    flowops=[uperf_schema.DisconnectFlowOp(type="disconnect")],
+                ),
+            ],
         )
-    ]
+    ],
 )
 
-sample_profile_expected = '''<?xml version='1.0' encoding='us-ascii'?>
+sample_profile_expected = """<?xml version='1.0' encoding='us-ascii'?>
 <profile name="test">
   <group nthreads="1">
     <transaction iterations="1">
@@ -65,37 +56,30 @@ sample_profile_expected = '''<?xml version='1.0' encoding='us-ascii'?>
     </transaction>
   </group>
 </profile>
-'''
+"""  # noqa: E501
+
 
 class ExamplePluginTest(unittest.TestCase):
     @staticmethod
     def test_serialization():
         plugin.test_object_serialization(simple_profile)
 
-        plugin.test_object_serialization(
-            uperf_plugin.UPerfServerParams(5)
-        )
+        plugin.test_object_serialization(uperf_plugin.UPerfServerParams(5))
 
-        plugin.test_object_serialization(
-            uperf_plugin.UPerfResults("test", {})
-        )
+        plugin.test_object_serialization(uperf_plugin.UPerfResults("test", {}))
 
-        plugin.test_object_serialization(
-            uperf_plugin.UPerfServerResults()
-        )
+        plugin.test_object_serialization(uperf_plugin.UPerfServerResults())
 
-        plugin.test_object_serialization(
-            uperf_plugin.UPerfError("Some error")
-        )
+        plugin.test_object_serialization(uperf_plugin.UPerfError("Some error"))
 
         plugin.test_object_serialization(
             uperf_plugin.UPerfServerError(1, "Not a real error")
         )
-    
+
     def test_profile_gen(self):
         uperf_plugin.clean_profile()
         uperf_plugin.write_profile(simple_profile)
-        with open(uperf_plugin.profile_path, 'r', encoding='us-ascii') as file:
+        with open(uperf_plugin.profile_path, "r", encoding="us-ascii") as file:
             generated_file = file.read()
         uperf_plugin.clean_profile()
         self.assertEqual(sample_profile_expected, generated_file)
@@ -113,7 +97,7 @@ class ExamplePluginTest(unittest.TestCase):
         # Test for error when port is in use
         # Make socket for port be in use.
         server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        server.bind(('127.0.0.1', 20000))
+        server.bind(("127.0.0.1", 20000))
         server.listen(8)
         server.setblocking(False)
         # Run the server, which should fail.
@@ -125,10 +109,15 @@ class ExamplePluginTest(unittest.TestCase):
         # --------------------------
         # Test the client failing due to no server
 
-        with contextlib.redirect_stdout(None): # Hide error messages
+        with contextlib.redirect_stdout(None):  # Hide error messages
             output_id, output_obj = uperf_plugin.run_uperf(simple_profile)
         self.assertEqual("error", output_id)
-        self.assertEqual(1, output_obj.error.count("TCP: Cannot connect to 127.0.0.1:20000 Connection refused"))
+        self.assertEqual(
+            1,
+            output_obj.error.count(
+                "TCP: Cannot connect to 127.0.0.1:20000 Connection refused"
+            ),
+        )
 
         # --------------------------
         # Test an actual working scenario
@@ -136,5 +125,5 @@ class ExamplePluginTest(unittest.TestCase):
         # TODO. This requires multiple processes.
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
